pyproject.toml
[project]
name = "chat-room-for-bitrix24"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn",
    "aiogram",
    "sqlalchemy",
    "alembic",
    "pydantic",
    "pydantic-settings",
    "python-jose",
    "passlib",
    "python-multipart",
    "jinja2",
    "aiosqlite",
    "asyncpg",
    "email-validator",
    "greenlet>=3.2.1",
    "bcrypt>=4.3.0",
]
